% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cancer.R
\docType{data}
\name{cancer}
\alias{cancer}
\title{Liver cancer data}
\format{
An object of class \code{data.frame} with 903 rows and 8 columns.
}
\usage{
data(cancer)
}
\description{
Survival in patients with liver cancer from the the Surveillance,
Epidemiology, and End Results (SEER) dataset.
}
\examples{
data(cancer)
head(cancer)
}
\keyword{datasets}
