% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-rcsplot.R
\name{print.rcsplot}
\alias{print.rcsplot}
\title{Print  rcsplot}
\usage{
\method{print}{rcsplot}(x, ...)
}
\arguments{
\item{x}{an object of rcsplot}

\item{...}{more arguments.}
}
\value{
No return value, called for drawing.
}
\description{
Print  rcsplot
}
\keyword{internal}
