% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RCS.R
\name{rcsplot}
\alias{rcsplot}
\title{Plot Restricted Cubic Splines Curves}
\usage{
rcsplot(
  data,
  outcome = NULL,
  time = NULL,
  exposure = NULL,
  covariates = NULL,
  positive = NULL,
  group = NULL,
  knots = c(0.05, 0.35, 0.65, 0.95),
  ref.value = "k1",
  conf.int = TRUE,
  conf.level = 0.95,
  conf.type = c("shape", "line"),
  pvalue = TRUE,
  pvalue.digits = 3,
  pvalue.position = c(0.02, 0.98),
  fontsize = 12,
  fontfamily = "serif",
  linesize = 0.25,
  linecolor = "#0072B5FF",
  alpha = 0.1,
  xbreaks = NULL,
  ybreaks = NULL,
  ...
)
}
\arguments{
\item{data}{a data frame contain the columns of outcome, time, exposure,
covariates, and group.}

\item{outcome}{the name of outcome variable in the data.}

\item{time}{the name of time variable in the data.}

\item{exposure}{the name of exposure variable in the data.}

\item{covariates}{the names of covariate variables in the data.}

\item{positive}{in which positive of outcome variable to make the comparison.
By default, positive is automatically defined. If outcome is a factor variable,
then positive is defined as the highest level. If outcome is a numerical
variable, then positive is defined as the largest value.}

\item{group}{the name of group variable in the data.}

\item{knots}{knots}

\item{ref.value}{ref.value}

\item{conf.int}{logical indicating whether or not to draw confidence interval.
Defaults to TRUE.}

\item{conf.level}{the confidence level to use for the confidence interval if
conf.int = TRUE. Must be strictly greater than 0 and less than 1. Defaults to
0.95, which corresponds to a 95 percent confidence interval.}

\item{conf.type}{conf.type}

\item{pvalue}{logical indicating whether or not to show P values,
include P for overall association and P for nonlinear, default TRUE.}

\item{pvalue.digits}{digits for P values, default 3.}

\item{pvalue.position}{position for P value, numeric vector of length two (x-axis and y-axis).}

\item{fontsize}{font size, default 12.}

\item{fontfamily}{font family, default 'serif'.}

\item{linesize}{line size, default 0.25.}

\item{linecolor}{line color, default '#0072B5FF'.}

\item{alpha}{alpha for the shape of confidence interval, default 0.1.}

\item{xbreaks}{breaks of x-axis.}

\item{ybreaks}{breaks of y-axis.}

\item{...}{further arguments.}
}
\value{
a ggplot2 object.
}
\description{
Plot Restricted Cubic Splines Curves
}
